requirements.txt
pymysql
pandas
sqlalchemy
clickhouse-driver
requests
prophet
matplotlib
python-dotenv